[ 172 ] organized lab 3, remove not used graphics.
