+++ 172/lab-3/report/report.tex
@@ -21,6 +21,8 @@
 \definecolor{customGray}{RGB}{128,128,128}
 \definecolor{Eblue}{rgb}{0.0, 0.18, 0.39}
 %==============Header & Footnote==============
+%
+\graphicspath{{figs/}}
 
 \pagestyle{fancy}
 \renewcommand{\headrulewidth}{0pt}
