--- 172/labs/3/TeX/Lab.bib
@@ -1,25 +1,0 @@
-@MISC {StackOverflow,
-  AUTHOR = "Sean",
-  TITLE = "What is a StackOverflowError?",
-  YEAR = "2008",
-  HOWPUBLISHED = {\url{https://stackoverflow.com/questions/214741/what-is-a-stackoverflowerror}}
-}
-
-@MISC {HeapStack,
-  AUTHOR = "Pankaj",
-  TITLE = "Java Heap Space vs. Stack – Memory Allocation in Java",
-  YEAR = "2017",
-  HOWPUBLISHED = {\url{http://www.journaldev.com/4098/java-heap-space-vs-stack-memory}}
-}
-
-@MISC {Wolfram,
-  AUTHOR = "WolframAlpha",
-  HOWPUBLISHED = {\url{https://www.wolframalpha.com/}}
-}
-
-@MISC{SJ,
-AUTHOR = "SJ",
-TITLE = "Introduction To Dynamic Programming: Fibonacci Series",
-YEAR = "2015",
-HOWPUBLISHED = {\url{http://algorithms.tutorialhorizon.com/introduction-to-dynamic-programming-fibonacci-series/}}
-}
